* 修正了一些很蠢的bug, 比如消息类型写错为Message::LogoutReqest; * 修正了BasicMessage的 readHeader(), writeHeader() 的错误用法, 为 PacketStream 类新增 downcaseTo() 方法, 以正确的执行 readHeader(), writeHeader(); * 修正了一些不太明显的小bug(复制粘贴时造成的错误);

+++ src/main/kvdb/stream/OutputStream.h
@@ -120,7 +120,7 @@ public:
 
     void writeInt32(int32_t value) {
         base_type::setInt32(value);
-        base_type::nextInt16();
+        base_type::nextInt32();
     }
 
     void writeInt64(int64_t value) {
